Fix goenv install for 1.3.1
There was a problem with the vercomp function that caused it to echo
more than once for some comparisons.  This would appear to be a side
effect of changing the stackoverflow function to echo instead of using
return value.

Comparing 1.2 and 1.3.1 is one of the cases that was echoing more than
once.  This then caused the test for `"$rtn" == "1"` to be false, which
meant that it was trying to use the old source URL.